chore: 클래스 병합(memberservice1, memberrepository1) 및 패키지 이동(querydlsconfig)
