+++ src/main/java/com/climingo/climingoApi/global/config/QuerydslConfig.java
@@ -1,4 +1,4 @@
-package com.climingo.climingoApi.config;
+package com.climingo.climingoApi.global.config;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import jakarta.persistence.EntityManager;
@@ -7,7 +7,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class QuerydslConfiguration {
+public class QuerydslConfig {
 
     @PersistenceContext
     private EntityManager entityManager;
