+++ src/main/java/com/climingo/climingoApi/member/domain/MemberRepository.java
@@ -2,6 +2,8 @@ package com.climingo.climingoApi.member.domain;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
@@ -12,4 +14,7 @@ public interface MemberRepository extends JpaRepository<Member, Long> {
     boolean existsByNickname(String nickname);
 
     Optional<Member> findByAuthIdAndProviderType(String authId, String providerType);
+
+    @Query("SELECT m FROM Member m JOIN FETCH m.records WHERE m.id = :memberId")
+    Member findMemberWithRecords(@Param("memberId") Long memberId);
 }
